Move errors into separate cog
